pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maestro"
version = "0.1.0"
description = "Unified Scrum Planner + Agent Orchestrator"
requires-python = ">=3.11"
dependencies = [
    # LangGraph / LangChain (planner)
    "langchain>=0.3",
    "langgraph>=0.2",
    "langchain-anthropic>=0.3",
    "langgraph-checkpoint-sqlite>=2.0",
    # HTTP / API
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "websockets>=13.0",
    # Config / env
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    # CLI / display
    "rich>=13.0",
    "click>=8.1",
    # Integrations
    "PyGithub>=2.0",
    "azure-devops>=7.0",
    "jira>=3.0",
    "atlassian-python-api>=3.0",
    "holidays>=0.40",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.3"]
google = ["langchain-google-genai>=2.0"]
pdf = ["pymupdf>=1.24"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-recording>=0.13",
    "ruff>=0.8",
    "pre-commit>=4.0",
    "syrupy>=4.0",
    "httpx",  # for FastAPI TestClient
]

[project.scripts]
maestro = "maestro.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/maestro"]

[tool.ruff]
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["maestro"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "eval: golden dataset evaluators",
    "vcr: contract tests with recorded API responses",
    "smoke: live API smoke tests (requires credentials)",
]
asyncio_mode = "auto"
